requirements.txt
redis==6.0.0
python-dotenv==1.0.1
Django==5.2
dj-rest-auth==4.0.1
django-allauth==0.54.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
psycopg2-binary
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
